Move search SQL query out of SearchHandler
